backend/requirements.txt
Django==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
mysqlclient==2.2.4
channels==4.0.0
channels-redis==4.2.0
python-dotenv==1.0.1
Pillow==10.2.0
